refactor(test): consolidate three TRON phase test files into tron-builders
PR 4 of test-suite-compaction. Merges three phase-segmented files
that all test pure TRON builder logic with the same fetch-stub
infrastructure:

  - test/tron-phase2-prepare.test.ts (387 LOC, 11 describe blocks
    covering base58/abi-encoding/native-send/token-send/claim-rewards)
  - test/tron-phase2b-stake-writes.test.ts (226 LOC, 3 blocks for
    Stake 2.0 freeze/unfreeze/withdraw-expire-unfreeze)
  - test/tron-phase2c-voting.test.ts (262 LOC, 2 blocks for SR
    listing + multi-SR vote)

→ test/tron-builders.test.ts (854 LOC, 16 blocks total).

Each describe block already manages its own `vi.stubGlobal("fetch",
...)` setup independently, so the merge is straightforward — just
unified imports + helpers (`directTxResponse` from phase2b, `fakeWitnessList`
from phase2c) at the top, then all describe blocks appended in order.

Coverage: 1116/1116 tests pass — exact baseline preserved.
File count: 95 → 93 (-2). LOC: -21 net (875 → 854).

The remaining gap from the surveyor's estimate is honest: each file's
preamble is similar but not redundant (different builders imported,
different action types tested), so the dedupe ceiling is small.
